# PERF: remove use of Python sets for interpolate