@simonjayhawkins 2020-06-12
some environments are failing with `TypeError: diff() got an unexpected keyword argument 'prepend'`