@simonjayhawkins 2020-07-24
I'll close this for now. still don't have access to desktop and running asvs on laptop is painful.